Proposal for issue priorities
@jerryjuwai, @imlazyone and @tjoelsson: Here is a proposal for issues priorities in Jira.

This should help the Team to know when to work on issues and Product Owners to define the issues’ priorities.

I purposely limited the number of types to understand better what needs to be done and when. I would actually love to get rid of “high” or “critical”, if you have any suggestions.